requirements.txt

flask
flask-sqlalchemy
flask-marshmallow
flask-restful
marshmallow-sqlalchemy
pymysql
python-dotenv
flask-cors
Flask-Migrate
flask-bcrypt